Fixed warnings (#183)
* Fixed warnings

* Fixed warnings

* Fixed DX12GraphicsContext::onSetSubData2D issue

* CI disk space saving

* ci

* ci

* ci

+++ include/LuminoCore/Base/StlHelper.hpp
@@ -221,7 +221,7 @@ public:
 		std::basic_streambuf<TChar, std::char_traits<TChar> >::setg(buffer, buffer, bufferEnd);
 	}
 
-	int length() const
+	size_t length() const
 	{
 		return std::basic_streambuf<TChar, std::char_traits<TChar> >::pptr() - m_begin;
 	}
